Added a confirmation popup for delete need and updated html to have a delete button on each need, calling this removeNeed() function in dashboard.ts

+++ ufund-ui/ufund-frontend/src/app/admin/pages/dashboard/dashboard.ts
@@ -47,6 +47,9 @@ export class Dashboard implements OnInit{
    * @param id if of need to be deleted
    */
   removeNeed(id: number): void {
+    if (!confirm('Are you sure you want to delete this need?')) {
+      return;
+    }
     this.needsService.deleteNeed(id).subscribe({
       next: () => {
         // Remove from local array so UI updates immediately
